aufgabe 1

+++ Praktikum 2/src/Token.java
@@ -16,6 +16,6 @@ public class Token {
     public int pos;  // position
     public int line; // line for error
     public int col;  // position for error
-    public int val;  // for numbers
+    public double val;  // for numbers
     public String str;  // for numbers and identifiers
 }
